Broke 2D, and 3D is not working yet.
Added example runner for the different testing files. Some internal changes
have broken the working 2D example. Temporarily whilst I try to make 3D
work, and improve the API to a point where things are fine and dandy.

Something isn't working in the 3D example, as the rendering tries to access
a vertex that is out of bounds. Not sure what is going on yet...