Clean Up & Added FunDeps to HasRenderFn
Improved the HasRenderFn typeclass a bit to have bidirectional functional
dependencies for the context. Since contexttype A can only refer to context B,
and vice versa. This lets me write functions ONCE and reuse them elsewhere and
if there are more context combinations later, this allows for a simple extension
without having to write any more specialised functions.

Added a function for running a raw context accepting function that might be
useful until more of the JSaddle API for the various drawing processes is
exposed or managed.