Attempt to clear up the typefamilies.
The typefamilies as they are currently being used are great for type safety. But
they become a little awkward when attempting to write one function to rule them
all. As I cannot convince the type checker that the 'c' is the same across the
board and that my type functions evaluate as I might expect.

This functionality should be simple enough to wrap up in the HasRenderFn
typeclass. The work continues.