New version! Yay.
Changed the dependency from `jsaddle-dom` to `ghcjs-dom` as the latter
is properly cached in nix.

Updated the various imports to match the use of `GHCJS.DOM` over
`JSDOM`.

Removed the `-dedupe` and `cpp-options` from the cabal file as these
options should be left to the end user of the library. Additionally
these flags were preventing the documentation from being compiled with
GHCJS, so there is no need for the `dontHaddock` override in the nix
expression.

Example project has been updated to suit.